Modification du type de retour des fonctions d'analyse fréquentielle
Il vaut mieux utiliser un double au lieu d'un unsigned int pour
stocker les fréquences de caractères.

+++ inc/c2.h
@@ -3,8 +3,8 @@
 
 #include "types.h"
 
-uint8 frequency_char(byte str[], uint32 str_length, byte c);
-uint16 frequency_analysis(byte str[], uint32 str_length);
+double frequency_char(byte str[], uint32 str_length, byte c);
+double frequency_analysis(byte str[], uint32 str_length);
 void c2(byte str[], uint32 str_length, byte **keys, uint32 nb_keys);
 
 #endif // C2_H
